broke route functions out into their own files
